+++ database/migrations/2022_01_19_110643_create_discount_students_table.php
@@ -15,8 +15,8 @@ class CreateDiscountStudentsTable extends Migration
     {
         Schema::create('discount_students', function (Blueprint $table) {
             $table->id();
-            $table->integer('assign_student_id');
-            $table->integer('fee_category_id')->nullable();
+            $table->unsignedBigInteger('assign_student_id');
+            $table->unsignedBigInteger('fee_category_id')->nullable();
             $table->double('discount')->nullable();
             $table->timestamps();
         });
