+++ database/migrations/2022_02_02_210524_create_employee_attendances_table.php
@@ -15,7 +15,7 @@ class CreateEmployeeAttendancesTable extends Migration
     {
         Schema::create('employee_attendances', function (Blueprint $table) {
             $table->id();
-            $table->integer('employee_id')->comment('employee_id = User id');
+            $table->unsignedBigInteger('employee_id')->comment('employee_id = User id');
             $table->date('date');
             $table->string('attendance_status')->comment('attendance_status = Present, Absent, Leave, Half Day');
             $table->timestamps();
