+++ database/migrations/2022_02_05_200352_create_account_employee_salaries_table.php
@@ -15,7 +15,7 @@ class CreateAccountEmployeeSalariesTable extends Migration
     {
         Schema::create('account_employee_salaries', function (Blueprint $table) {
             $table->id();
-            $table->integer('employee_id')->comment('employee_id = user_id');
+            $table->unsignedBigInteger('employee_id')->comment('employee_id = user_id');
             $table->string('date')->nullable();
             $table->double('amount')->nullable();
             $table->timestamps();
