+++ database/migrations/2022_11_25_173557_create_student_attendances_table.php
@@ -15,12 +15,12 @@ class CreateStudentAttendancesTable extends Migration
     {
         Schema::create('student_attendances', function (Blueprint $table) {
             $table->id();
-            $table->integer('student_id')->comment('student_id = User id');
-            $table->integer('roll')->nullable();
-            $table->integer('year_id')->nullable();
-            $table->integer('shift_id')->nullable();
-            $table->integer('section_id')->nullable();
-            $table->integer('group_id')->nullable();
+            $table->unsignedBigInteger('student_id')->comment('student_id = User id');
+            $table->unsignedBigInteger('roll')->nullable();
+            $table->unsignedBigInteger('year_id')->nullable();
+            $table->unsignedBigInteger('shift_id')->nullable();
+            $table->unsignedBigInteger('section_id')->nullable();
+            $table->unsignedBigInteger('group_id')->nullable();
             $table->date('date');
             $table->string('attendance_status')->comment('attendance_status = Present, Absent, Leave, Half Day');
             $table->timestamps();
