+++ database/migrations/2023_05_13_030125_create_student_payments_table.php
@@ -18,7 +18,7 @@ class CreateStudentPaymentsTable extends Migration
             $table->unsignedBigInteger('invoice_id');
             $table->foreign('invoice_id')->references('id')->on('invoices');
             $table->unsignedBigInteger('student_id');
-            $table->Integer('amount');
+            $table->double('amount');
             $table->date('payment_date');
             $table->timestamps();
         });
